fix(buttons): fixed imports and exports (EC-4)
