--- src/app/components/Buttons/index.ts
@@ -1,1 +1,0 @@
-export { PrimaryButton, IconButton } from "./Buttons";
