+++ src/app/components/IconButton/IconButton.types.ts
@@ -1,7 +1,3 @@
-export type PrimaryButtonProps = {
-  buttonLabel: string;
-};
-
 export type IconButtonProps = {
   classNames?: string;
   iconPath: string;
