+++ src/app/components/IconButton/index.ts
@@ -1,0 +1,3 @@
+import IconButton from "./IconButton";
+
+export default IconButton;
