+++ src/app/components/PrimaryButton/PrimaryButton.types.ts
@@ -1,0 +1,3 @@
+export type PrimaryButtonProps = {
+  buttonLabel: string;
+};
