+++ src/app/components/PrimaryButton/index.ts
@@ -1,0 +1,3 @@
+import PrimaryButton from "./PrimaryButton";
+
+export default PrimaryButton;
